Classe password

+++ pass_gen/database_interact.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 
 import sqlite3
+from typing import Optional
 import pyperclip
 
 try:
@@ -9,6 +10,7 @@ except:
     conn = sqlite3.connect('passwords/database.sqlite')
 finally:
     curs = conn.cursor()
+
 
 def table_creation():
     try:
@@ -21,14 +23,17 @@ def table_creation():
     except sqlite3.OperationalError:
         return None
 
+
 def get_info():
     curs.execute("SELECT * FROM user_passwords")
     return curs.fetchall()
 
-def insert_info(user, target_site, user_password):
+
+def insert_info(user: str, target_site: str, user_password: str):
     with conn:
         curs.execute("INSERT INTO user_passwords VALUES (:username, :site, :password)",
-         {'username':user, 'site':target_site, 'password':user_password})
+                     {'username': user, 'site': target_site, 'password': user_password})
+
 
 def see_save():
     info = get_info()
@@ -36,29 +41,31 @@ def see_save():
     for i in lista:
         print(i)
 
+
 def get_passwords():
     info = get_info()
     lista = [f'{n} - {i[0]} para o site {i[1]}' for n, i in enumerate(info, 1)]
     for i in lista:
         print(i)
-    
+
     index = int(input("Qual é a palavra passe que quer? "))
 
     try:
-        password = info[index-1][2]
+        password = info[index - 1][2]
         pyperclip.copy(password)
         print("A palavra pass foi adicionada à área de transferência.")
     except:
         print('Fiz algo não muito bom!')
 
-def delete_one(index):
-    info = get_info()
-    target_site = info[index-1][1]
 
-    with conn:
-        curs.execute("DELETE from user_passwords WHERE site = :site",
-        {'site':target_site})
+def delete_info(index: Optional[int] = None, one: bool = False, everything: bool = False):
+    if one:
+        info = get_info()
+        target_site = info[index - 1][1]
 
-def delete_all():
-    with conn:
-        curs.execute("DELETE from user_passwords")
+        with conn:
+            curs.execute("DELETE from user_passwords WHERE site = :site",
+                         {'site': target_site})
+    if everything:
+        with conn:
+            curs.execute("DELETE from user_passwords")
